fix: update resume position type to double (#443)
* fix: update resume position type to double

* chore: no need to convert

+++ ios/JSProps/Source.swift
@@ -36,7 +36,7 @@ struct Source: SuperCodable {
     let dvrSeekForwardInterval: Int?
     let dvrSeekBackwardInterval: Int?
     var plugins: JSPlugins?
-    let resumePosition: Int?
+    let resumePosition: Double?
 }
